# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.DecryptAuditLog:
    properties:
      audit_log_id:
        type: string
      client_ip:
        type: string
      decrypted_at:
        type: string
      fail_reason:
        type: string
      success:
        type: boolean
      tenant_id:
        type: string
      user_id:
        type: string
    type: object
  api.DecryptAuditLogsResponse:
    properties:
      encrypted_ref:
        type: string
      logs:
        items:
          $ref: '#/definitions/api.DecryptAuditLog'
        type: array
      total_count:
        type: integer
    type: object
  api.DecryptRequest:
    properties:
      content_type:
        description: prompt 或 output
        type: string
      encrypted_ref:
        description: 加密内容引用
        type: string
      trace_id:
        description: 可选的 trace_id
        type: string
    required:
    - content_type
    - encrypted_ref
    type: object
  api.DecryptResponse:
    properties:
      audit_log_id:
        description: 审计日志 ID
        type: string
      content:
        description: 解密后的内容
        type: string
      content_type:
        description: 内容类型
        type: string
      decrypted_at:
        description: 解密时间
        type: string
    type: object
...
host: localhost:8006
info:
  contact:
    email: support@aitrace.cc
    name: AI-Trace Support
    url: https://aitrace.cc/support
  description: |-
    AI-Trace - Enterprise AI Decision Audit & Compliance Platform

    AI-Trace 提供零代码侵入的 AI 调用追踪、存证和审计能力。
    支持 OpenAI、Claude、Ollama 等多种模型，用户可透传自己的 API Key，服务不存储任何敏感信息。

    ## 特性
    - **零侵入集成**: 只需替换 API endpoint，无需修改代码
    - **API Key 透传**: 用户的 API Key 直接透传到上游，不存储
    - **自定义代理**: 支持用户指定自己的代理服务器
    - **多级存证**: internal(内部审计) / compliance(合规) / legal(法律效力)
    - **最小披露**: 基于 Merkle 树的选择性披露证明
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://aitrace.cc/terms
  title: AI-Trace API
  version: 0.2.0
paths:
...
  /decrypt:
    post:
      consumes:
      - application/json
      description: |-
        解密已加密存储的 prompt 或 output 内容

        ## 权限要求
        - 只有内容所有者或授权方可解密
        - 所有解密操作会被记录到审计日志
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 解密请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.DecryptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 解密结果
          schema:
            $ref: '#/definitions/api.DecryptResponse'
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 内容未找到
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 解密内容
      tags:
      - Decrypt
  /decrypt/audit:
    get:
      description: 获取指定内容的解密审计历史
      parameters:
      - description: 加密内容引用
        in: query
        name: encrypted_ref
        required: true
        type: string
      - default: 20
        description: 返回数量限制
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 审计日志
          schema:
            $ref: '#/definitions/api.DecryptAuditLogsResponse'
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取解密审计日志
      tags:
      - Decrypt
...
securityDefinitions:
  ApiKeyAuth:
    description: AI-Trace 平台 API Key（用于认证）
    in: header
    name: X-API-Key
    type: apiKey
  UpstreamApiKey:
    description: 用户的上游 API Key（OpenAI/Claude，透传不存储）
    in: header
    name: X-Upstream-API-Key
    type: apiKey
swagger: "2.0"